Implement getStateSummary for Node at base
So a Node can decide if it wants to offer a summary, or just a single
state.
